algorithms/medium/non-overlapping-intervals-435/src/main/java/ru/saidgadjiev/leetcode/algorithms/_435/Solution.java
package ru.saidgadjiev.leetcode.algorithms._435;

import java.util.Arrays;
import java.util.Comparator;

/**
 * https://leetcode.com/problems/non-overlapping-intervals
 */
public class Solution {

    public int eraseOverlapIntervals(int[][] intervals) {
        Arrays.sort(intervals, Comparator.comparingInt(o -> o[1]));

        int prevEnd = intervals[0][1];
        int result = 0;

        for (int i = 1; i < intervals.length; i++) {
            if (intervals[i][0] >= prevEnd) {
                prevEnd = intervals[i][1];
            } else {
                ++result;
            }
        }

        return result;
    }
}
